power/test_launcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Power API</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body/>

  <script type="application/javascript">
  // It is a requirement that power_basics.html is in the directory path that is noted below. 
  // If the file is not present then the test will fail silently

  SimpleTest.waitForExplicitFinish();
  var url = "http://example.org/tests/dom/power/test/power_basics.html";
  
  SpecialPowers.addPermission("power", true, url);

  var iframe = document.createElement("iframe");
  iframe.mozApp = "http://example.org/manifest.webapp";
  iframe.src = url;
  iframe.setAttribute("id", "ifrm");
  iframe.onload = loaded();
  document.body.appendChild(iframe);

  function loaded() {

    window.addEventListener("message", function onMessage(event) {
       if (event.data == "finished") {
         window.removeEventListener("message", onMessage, false);
         SimpleTest.finish();
       } else {
         dump(JSON.stringify(event.data));
         ok(event.data.result, event.data.msg);
       }
           
     }, false);

  }
  </script>
</head>

</html>
